requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
Django==1.11.4
django-debug-toolbar==1.8
django-extensions==1.9.1
django-paypal==0.4.1
idna==2.6
ipdb==0.10.3
ipython-genutils==0.2.0
jedi==0.10.2
numpy==1.13.1
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
pytest==3.2.5
pytest-sugar==0.8.0
pytest-django==3.1.2
django-model-utils==3.0.0
opencv-python==3.3.0.10 
